+++ protos/vega/commands/v1/transaction.pb.go
@@ -20,6 +20,59 @@ const (
 	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
 )
 
+// The current supported version of the transaction inside the network.
+type TxVersion int32
+
+const (
+	TxVersion_TX_VERSION_UNSPECIFIED TxVersion = 0
+	// This version requires the Proof-of-work added to the transaction.
+	TxVersion_TX_VERSION_V2 TxVersion = 2
+	// This version requires the chain ID to be append in front of the input data
+	// byte, with a `\0` delimiter.
+	TxVersion_TX_VERSION_V3 TxVersion = 3
+)
+
+// Enum value maps for TxVersion.
+var (
+	TxVersion_name = map[int32]string{
+		0: "TX_VERSION_UNSPECIFIED",
+		2: "TX_VERSION_V2",
+		3: "TX_VERSION_V3",
+	}
+	TxVersion_value = map[string]int32{
+		"TX_VERSION_UNSPECIFIED": 0,
+		"TX_VERSION_V2":          2,
+		"TX_VERSION_V3":          3,
+	}
+)
+
+func (x TxVersion) Enum() *TxVersion {
+	p := new(TxVersion)
+	*p = x
+	return p
+}
+
+func (x TxVersion) String() string {
+	return protoimpl.X.EnumStringOf(x.Descriptor(), protoreflect.EnumNumber(x))
+}
+
+func (TxVersion) Descriptor() protoreflect.EnumDescriptor {
+	return file_vega_commands_v1_transaction_proto_enumTypes[0].Descriptor()
+}
+
+func (TxVersion) Type() protoreflect.EnumType {
+	return &file_vega_commands_v1_transaction_proto_enumTypes[0]
+}
+
+func (x TxVersion) Number() protoreflect.EnumNumber {
+	return protoreflect.EnumNumber(x)
+}
+
+// Deprecated: Use TxVersion.Descriptor instead.
+func (TxVersion) EnumDescriptor() ([]byte, []int) {
+	return file_vega_commands_v1_transaction_proto_rawDescGZIP(), []int{0}
+}
+
 type InputData struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
@@ -437,7 +490,9 @@ type Transaction struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	// One of the set of Vega commands (proto marshalled).
+	// One of the set of Vega commands.
+	// These bytes are should be built as follow:
+	//   chain_id_as_bytes + \0 character as delimiter + proto_marshalled_command
 	InputData []byte `protobuf:"bytes,1,opt,name=input_data,json=inputData,proto3" json:"input_data,omitempty"`
 	// The signature of the inputData.
 	Signature *Signature `protobuf:"bytes,2,opt,name=signature,proto3" json:"signature,omitempty"`
@@ -450,7 +505,7 @@ type Transaction struct {
 	From isTransaction_From `protobuf_oneof:"from"`
 	// A version of the transaction, to be used in the future in case changes are implemented
 	// to the Transaction format.
-	Version uint32 `protobuf:"varint,2000,opt,name=version,proto3" json:"version,omitempty"`
+	Version TxVersion `protobuf:"varint,2000,opt,name=version,proto3,enum=vega.commands.v1.TxVersion" json:"version,omitempty"`
 	// Proof of work contains the random transaction id used by the client and the nonce.
 	Pow *ProofOfWork `protobuf:"bytes,3000,opt,name=pow,proto3" json:"pow,omitempty"`
 }
@@ -522,11 +577,11 @@ func (x *Transaction) GetPubKey() string {
 	return ""
 }
 
-func (x *Transaction) GetVersion() uint32 {
+func (x *Transaction) GetVersion() TxVersion {
 	if x != nil {
 		return x.Version
 	}
-	return 0
+	return TxVersion_TX_VERSION_UNSPECIFIED
 }
 
 func (x *Transaction) GetPow() *ProofOfWork {
@@ -559,9 +614,8 @@ type ProofOfWork struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	Tid          string `protobuf:"bytes,1,opt,name=tid,proto3" json:"tid,omitempty"`
-	Nonce        uint64 `protobuf:"varint,2,opt,name=nonce,proto3" json:"nonce,omitempty"`
-	HashFunction string `protobuf:"bytes,3,opt,name=hash_function,json=hashFunction,proto3" json:"hash_function,omitempty"`
+	Tid   string `protobuf:"bytes,1,opt,name=tid,proto3" json:"tid,omitempty"`
+	Nonce uint64 `protobuf:"varint,2,opt,name=nonce,proto3" json:"nonce,omitempty"`
 }
 
 func (x *ProofOfWork) Reset() {
@@ -608,13 +662,6 @@ func (x *ProofOfWork) GetNonce() uint64 {
 		return x.Nonce
 	}
 	return 0
-}
-
-func (x *ProofOfWork) GetHashFunction() string {
-	if x != nil {
-		return x.HashFunction
-	}
-	return ""
 }
 
 var File_vega_commands_v1_transaction_proto protoreflect.FileDescriptor
@@ -767,7 +814,7 @@ var file_vega_commands_v1_transaction_proto_rawDesc = []byte{
 	0x61, 0x53, 0x75, 0x62, 0x6d, 0x69, 0x73, 0x73, 0x69, 0x6f, 0x6e, 0x48, 0x00, 0x52, 0x14, 0x6f,
 	0x72, 0x61, 0x63, 0x6c, 0x65, 0x44, 0x61, 0x74, 0x61, 0x53, 0x75, 0x62, 0x6d, 0x69, 0x73, 0x73,
 	0x69, 0x6f, 0x6e, 0x42, 0x09, 0x0a, 0x07, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x4a, 0x06,
-	0x08, 0xa1, 0x1f, 0x10, 0xa2, 0x1f, 0x22, 0xf5, 0x01, 0x0a, 0x0b, 0x54, 0x72, 0x61, 0x6e, 0x73,
+	0x08, 0xa1, 0x1f, 0x10, 0xa2, 0x1f, 0x22, 0x92, 0x02, 0x0a, 0x0b, 0x54, 0x72, 0x61, 0x6e, 0x73,
 	0x61, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x1d, 0x0a, 0x0a, 0x69, 0x6e, 0x70, 0x75, 0x74, 0x5f,
 	0x64, 0x61, 0x74, 0x61, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x09, 0x69, 0x6e, 0x70, 0x75,
 	0x74, 0x44, 0x61, 0x74, 0x61, 0x12, 0x39, 0x0a, 0x09, 0x73, 0x69, 0x67, 0x6e, 0x61, 0x74, 0x75,
@@ -777,22 +824,27 @@ var file_vega_commands_v1_transaction_proto_rawDesc = []byte{
 	0x12, 0x1b, 0x0a, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x18, 0xe9, 0x07, 0x20, 0x01,
 	0x28, 0x09, 0x48, 0x00, 0x52, 0x07, 0x61, 0x64, 0x64, 0x72, 0x65, 0x73, 0x73, 0x12, 0x1a, 0x0a,
 	0x07, 0x70, 0x75, 0x62, 0x5f, 0x6b, 0x65, 0x79, 0x18, 0xea, 0x07, 0x20, 0x01, 0x28, 0x09, 0x48,
-	0x00, 0x52, 0x06, 0x70, 0x75, 0x62, 0x4b, 0x65, 0x79, 0x12, 0x19, 0x0a, 0x07, 0x76, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x18, 0xd0, 0x0f, 0x20, 0x01, 0x28, 0x0d, 0x52, 0x07, 0x76, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x12, 0x30, 0x0a, 0x03, 0x70, 0x6f, 0x77, 0x18, 0xb8, 0x17, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1d, 0x2e, 0x76, 0x65, 0x67, 0x61, 0x2e, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e,
-	0x64, 0x73, 0x2e, 0x76, 0x31, 0x2e, 0x50, 0x72, 0x6f, 0x6f, 0x66, 0x4f, 0x66, 0x57, 0x6f, 0x72,
-	0x6b, 0x52, 0x03, 0x70, 0x6f, 0x77, 0x42, 0x06, 0x0a, 0x04, 0x66, 0x72, 0x6f, 0x6d, 0x22, 0x5a,
-	0x0a, 0x0b, 0x50, 0x72, 0x6f, 0x6f, 0x66, 0x4f, 0x66, 0x57, 0x6f, 0x72, 0x6b, 0x12, 0x10, 0x0a,
-	0x03, 0x74, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x74, 0x69, 0x64, 0x12,
-	0x14, 0x0a, 0x05, 0x6e, 0x6f, 0x6e, 0x63, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x05,
-	0x6e, 0x6f, 0x6e, 0x63, 0x65, 0x12, 0x23, 0x0a, 0x0d, 0x68, 0x61, 0x73, 0x68, 0x5f, 0x66, 0x75,
-	0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x68, 0x61,
-	0x73, 0x68, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x42, 0x33, 0x5a, 0x31, 0x63, 0x6f,
-	0x64, 0x65, 0x2e, 0x76, 0x65, 0x67, 0x61, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2e,
-	0x69, 0x6f, 0x2f, 0x76, 0x65, 0x67, 0x61, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f, 0x76,
-	0x65, 0x67, 0x61, 0x2f, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x73, 0x2f, 0x76, 0x31, 0x62,
-	0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x00, 0x52, 0x06, 0x70, 0x75, 0x62, 0x4b, 0x65, 0x79, 0x12, 0x36, 0x0a, 0x07, 0x76, 0x65, 0x72,
+	0x73, 0x69, 0x6f, 0x6e, 0x18, 0xd0, 0x0f, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x1b, 0x2e, 0x76, 0x65,
+	0x67, 0x61, 0x2e, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x73, 0x2e, 0x76, 0x31, 0x2e, 0x54,
+	0x78, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x52, 0x07, 0x76, 0x65, 0x72, 0x73, 0x69, 0x6f,
+	0x6e, 0x12, 0x30, 0x0a, 0x03, 0x70, 0x6f, 0x77, 0x18, 0xb8, 0x17, 0x20, 0x01, 0x28, 0x0b, 0x32,
+	0x1d, 0x2e, 0x76, 0x65, 0x67, 0x61, 0x2e, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x73, 0x2e,
+	0x76, 0x31, 0x2e, 0x50, 0x72, 0x6f, 0x6f, 0x66, 0x4f, 0x66, 0x57, 0x6f, 0x72, 0x6b, 0x52, 0x03,
+	0x70, 0x6f, 0x77, 0x42, 0x06, 0x0a, 0x04, 0x66, 0x72, 0x6f, 0x6d, 0x22, 0x35, 0x0a, 0x0b, 0x50,
+	0x72, 0x6f, 0x6f, 0x66, 0x4f, 0x66, 0x57, 0x6f, 0x72, 0x6b, 0x12, 0x10, 0x0a, 0x03, 0x74, 0x69,
+	0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x74, 0x69, 0x64, 0x12, 0x14, 0x0a, 0x05,
+	0x6e, 0x6f, 0x6e, 0x63, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x05, 0x6e, 0x6f, 0x6e,
+	0x63, 0x65, 0x2a, 0x53, 0x0a, 0x09, 0x54, 0x78, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12,
+	0x1a, 0x0a, 0x16, 0x54, 0x58, 0x5f, 0x56, 0x45, 0x52, 0x53, 0x49, 0x4f, 0x4e, 0x5f, 0x55, 0x4e,
+	0x53, 0x50, 0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x11, 0x0a, 0x0d, 0x54,
+	0x58, 0x5f, 0x56, 0x45, 0x52, 0x53, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x32, 0x10, 0x02, 0x12, 0x11,
+	0x0a, 0x0d, 0x54, 0x58, 0x5f, 0x56, 0x45, 0x52, 0x53, 0x49, 0x4f, 0x4e, 0x5f, 0x56, 0x33, 0x10,
+	0x03, 0x22, 0x04, 0x08, 0x01, 0x10, 0x01, 0x42, 0x33, 0x5a, 0x31, 0x63, 0x6f, 0x64, 0x65, 0x2e,
+	0x76, 0x65, 0x67, 0x61, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x63, 0x6f, 0x6c, 0x2e, 0x69, 0x6f, 0x2f,
+	0x76, 0x65, 0x67, 0x61, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x2f, 0x76, 0x65, 0x67, 0x61,
+	0x2f, 0x63, 0x6f, 0x6d, 0x6d, 0x61, 0x6e, 0x64, 0x73, 0x2f, 0x76, 0x31, 0x62, 0x06, 0x70, 0x72,
+	0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -807,67 +859,70 @@ func file_vega_commands_v1_transaction_proto_rawDescGZIP() []byte {
 	return file_vega_commands_v1_transaction_proto_rawDescData
 }
 
+var file_vega_commands_v1_transaction_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
 var file_vega_commands_v1_transaction_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
 var file_vega_commands_v1_transaction_proto_goTypes = []interface{}{
-	(*InputData)(nil),                      // 0: vega.commands.v1.InputData
-	(*Transaction)(nil),                    // 1: vega.commands.v1.Transaction
-	(*ProofOfWork)(nil),                    // 2: vega.commands.v1.ProofOfWork
-	(*OrderSubmission)(nil),                // 3: vega.commands.v1.OrderSubmission
-	(*OrderCancellation)(nil),              // 4: vega.commands.v1.OrderCancellation
-	(*OrderAmendment)(nil),                 // 5: vega.commands.v1.OrderAmendment
-	(*WithdrawSubmission)(nil),             // 6: vega.commands.v1.WithdrawSubmission
-	(*ProposalSubmission)(nil),             // 7: vega.commands.v1.ProposalSubmission
-	(*VoteSubmission)(nil),                 // 8: vega.commands.v1.VoteSubmission
-	(*LiquidityProvisionSubmission)(nil),   // 9: vega.commands.v1.LiquidityProvisionSubmission
-	(*DelegateSubmission)(nil),             // 10: vega.commands.v1.DelegateSubmission
-	(*UndelegateSubmission)(nil),           // 11: vega.commands.v1.UndelegateSubmission
-	(*LiquidityProvisionCancellation)(nil), // 12: vega.commands.v1.LiquidityProvisionCancellation
-	(*LiquidityProvisionAmendment)(nil),    // 13: vega.commands.v1.LiquidityProvisionAmendment
-	(*Transfer)(nil),                       // 14: vega.commands.v1.Transfer
-	(*CancelTransfer)(nil),                 // 15: vega.commands.v1.CancelTransfer
-	(*AnnounceNode)(nil),                   // 16: vega.commands.v1.AnnounceNode
-	(*NodeVote)(nil),                       // 17: vega.commands.v1.NodeVote
-	(*NodeSignature)(nil),                  // 18: vega.commands.v1.NodeSignature
-	(*ChainEvent)(nil),                     // 19: vega.commands.v1.ChainEvent
-	(*KeyRotateSubmission)(nil),            // 20: vega.commands.v1.KeyRotateSubmission
-	(*StateVariableProposal)(nil),          // 21: vega.commands.v1.StateVariableProposal
-	(*ValidatorHeartbeat)(nil),             // 22: vega.commands.v1.ValidatorHeartbeat
-	(*EthereumKeyRotateSubmission)(nil),    // 23: vega.commands.v1.EthereumKeyRotateSubmission
-	(*ProtocolUpgradeProposal)(nil),        // 24: vega.commands.v1.ProtocolUpgradeProposal
-	(*OracleDataSubmission)(nil),           // 25: vega.commands.v1.OracleDataSubmission
-	(*Signature)(nil),                      // 26: vega.commands.v1.Signature
+	(TxVersion)(0),                         // 0: vega.commands.v1.TxVersion
+	(*InputData)(nil),                      // 1: vega.commands.v1.InputData
+	(*Transaction)(nil),                    // 2: vega.commands.v1.Transaction
+	(*ProofOfWork)(nil),                    // 3: vega.commands.v1.ProofOfWork
+	(*OrderSubmission)(nil),                // 4: vega.commands.v1.OrderSubmission
+	(*OrderCancellation)(nil),              // 5: vega.commands.v1.OrderCancellation
+	(*OrderAmendment)(nil),                 // 6: vega.commands.v1.OrderAmendment
+	(*WithdrawSubmission)(nil),             // 7: vega.commands.v1.WithdrawSubmission
+	(*ProposalSubmission)(nil),             // 8: vega.commands.v1.ProposalSubmission
+	(*VoteSubmission)(nil),                 // 9: vega.commands.v1.VoteSubmission
+	(*LiquidityProvisionSubmission)(nil),   // 10: vega.commands.v1.LiquidityProvisionSubmission
+	(*DelegateSubmission)(nil),             // 11: vega.commands.v1.DelegateSubmission
+	(*UndelegateSubmission)(nil),           // 12: vega.commands.v1.UndelegateSubmission
+	(*LiquidityProvisionCancellation)(nil), // 13: vega.commands.v1.LiquidityProvisionCancellation
+	(*LiquidityProvisionAmendment)(nil),    // 14: vega.commands.v1.LiquidityProvisionAmendment
+	(*Transfer)(nil),                       // 15: vega.commands.v1.Transfer
+	(*CancelTransfer)(nil),                 // 16: vega.commands.v1.CancelTransfer
+	(*AnnounceNode)(nil),                   // 17: vega.commands.v1.AnnounceNode
+	(*NodeVote)(nil),                       // 18: vega.commands.v1.NodeVote
+	(*NodeSignature)(nil),                  // 19: vega.commands.v1.NodeSignature
+	(*ChainEvent)(nil),                     // 20: vega.commands.v1.ChainEvent
+	(*KeyRotateSubmission)(nil),            // 21: vega.commands.v1.KeyRotateSubmission
+	(*StateVariableProposal)(nil),          // 22: vega.commands.v1.StateVariableProposal
+	(*ValidatorHeartbeat)(nil),             // 23: vega.commands.v1.ValidatorHeartbeat
+	(*EthereumKeyRotateSubmission)(nil),    // 24: vega.commands.v1.EthereumKeyRotateSubmission
+	(*ProtocolUpgradeProposal)(nil),        // 25: vega.commands.v1.ProtocolUpgradeProposal
+	(*OracleDataSubmission)(nil),           // 26: vega.commands.v1.OracleDataSubmission
+	(*Signature)(nil),                      // 27: vega.commands.v1.Signature
 }
 var file_vega_commands_v1_transaction_proto_depIdxs = []int32{
-	3,  // 0: vega.commands.v1.InputData.order_submission:type_name -> vega.commands.v1.OrderSubmission
-	4,  // 1: vega.commands.v1.InputData.order_cancellation:type_name -> vega.commands.v1.OrderCancellation
-	5,  // 2: vega.commands.v1.InputData.order_amendment:type_name -> vega.commands.v1.OrderAmendment
-	6,  // 3: vega.commands.v1.InputData.withdraw_submission:type_name -> vega.commands.v1.WithdrawSubmission
-	7,  // 4: vega.commands.v1.InputData.proposal_submission:type_name -> vega.commands.v1.ProposalSubmission
-	8,  // 5: vega.commands.v1.InputData.vote_submission:type_name -> vega.commands.v1.VoteSubmission
-	9,  // 6: vega.commands.v1.InputData.liquidity_provision_submission:type_name -> vega.commands.v1.LiquidityProvisionSubmission
-	10, // 7: vega.commands.v1.InputData.delegate_submission:type_name -> vega.commands.v1.DelegateSubmission
-	11, // 8: vega.commands.v1.InputData.undelegate_submission:type_name -> vega.commands.v1.UndelegateSubmission
-	12, // 9: vega.commands.v1.InputData.liquidity_provision_cancellation:type_name -> vega.commands.v1.LiquidityProvisionCancellation
-	13, // 10: vega.commands.v1.InputData.liquidity_provision_amendment:type_name -> vega.commands.v1.LiquidityProvisionAmendment
-	14, // 11: vega.commands.v1.InputData.transfer:type_name -> vega.commands.v1.Transfer
-	15, // 12: vega.commands.v1.InputData.cancel_transfer:type_name -> vega.commands.v1.CancelTransfer
-	16, // 13: vega.commands.v1.InputData.announce_node:type_name -> vega.commands.v1.AnnounceNode
-	17, // 14: vega.commands.v1.InputData.node_vote:type_name -> vega.commands.v1.NodeVote
-	18, // 15: vega.commands.v1.InputData.node_signature:type_name -> vega.commands.v1.NodeSignature
-	19, // 16: vega.commands.v1.InputData.chain_event:type_name -> vega.commands.v1.ChainEvent
-	20, // 17: vega.commands.v1.InputData.key_rotate_submission:type_name -> vega.commands.v1.KeyRotateSubmission
-	21, // 18: vega.commands.v1.InputData.state_variable_proposal:type_name -> vega.commands.v1.StateVariableProposal
-	22, // 19: vega.commands.v1.InputData.validator_heartbeat:type_name -> vega.commands.v1.ValidatorHeartbeat
-	23, // 20: vega.commands.v1.InputData.ethereum_key_rotate_submission:type_name -> vega.commands.v1.EthereumKeyRotateSubmission
-	24, // 21: vega.commands.v1.InputData.protocol_upgrade_proposal:type_name -> vega.commands.v1.ProtocolUpgradeProposal
-	25, // 22: vega.commands.v1.InputData.oracle_data_submission:type_name -> vega.commands.v1.OracleDataSubmission
-	26, // 23: vega.commands.v1.Transaction.signature:type_name -> vega.commands.v1.Signature
-	2,  // 24: vega.commands.v1.Transaction.pow:type_name -> vega.commands.v1.ProofOfWork
-	25, // [25:25] is the sub-list for method output_type
-	25, // [25:25] is the sub-list for method input_type
-	25, // [25:25] is the sub-list for extension type_name
-	25, // [25:25] is the sub-list for extension extendee
-	0,  // [0:25] is the sub-list for field type_name
+	4,  // 0: vega.commands.v1.InputData.order_submission:type_name -> vega.commands.v1.OrderSubmission
+	5,  // 1: vega.commands.v1.InputData.order_cancellation:type_name -> vega.commands.v1.OrderCancellation
+	6,  // 2: vega.commands.v1.InputData.order_amendment:type_name -> vega.commands.v1.OrderAmendment
+	7,  // 3: vega.commands.v1.InputData.withdraw_submission:type_name -> vega.commands.v1.WithdrawSubmission
+	8,  // 4: vega.commands.v1.InputData.proposal_submission:type_name -> vega.commands.v1.ProposalSubmission
+	9,  // 5: vega.commands.v1.InputData.vote_submission:type_name -> vega.commands.v1.VoteSubmission
+	10, // 6: vega.commands.v1.InputData.liquidity_provision_submission:type_name -> vega.commands.v1.LiquidityProvisionSubmission
+	11, // 7: vega.commands.v1.InputData.delegate_submission:type_name -> vega.commands.v1.DelegateSubmission
+	12, // 8: vega.commands.v1.InputData.undelegate_submission:type_name -> vega.commands.v1.UndelegateSubmission
+	13, // 9: vega.commands.v1.InputData.liquidity_provision_cancellation:type_name -> vega.commands.v1.LiquidityProvisionCancellation
+	14, // 10: vega.commands.v1.InputData.liquidity_provision_amendment:type_name -> vega.commands.v1.LiquidityProvisionAmendment
+	15, // 11: vega.commands.v1.InputData.transfer:type_name -> vega.commands.v1.Transfer
+	16, // 12: vega.commands.v1.InputData.cancel_transfer:type_name -> vega.commands.v1.CancelTransfer
+	17, // 13: vega.commands.v1.InputData.announce_node:type_name -> vega.commands.v1.AnnounceNode
+	18, // 14: vega.commands.v1.InputData.node_vote:type_name -> vega.commands.v1.NodeVote
+	19, // 15: vega.commands.v1.InputData.node_signature:type_name -> vega.commands.v1.NodeSignature
+	20, // 16: vega.commands.v1.InputData.chain_event:type_name -> vega.commands.v1.ChainEvent
+	21, // 17: vega.commands.v1.InputData.key_rotate_submission:type_name -> vega.commands.v1.KeyRotateSubmission
+	22, // 18: vega.commands.v1.InputData.state_variable_proposal:type_name -> vega.commands.v1.StateVariableProposal
+	23, // 19: vega.commands.v1.InputData.validator_heartbeat:type_name -> vega.commands.v1.ValidatorHeartbeat
+	24, // 20: vega.commands.v1.InputData.ethereum_key_rotate_submission:type_name -> vega.commands.v1.EthereumKeyRotateSubmission
+	25, // 21: vega.commands.v1.InputData.protocol_upgrade_proposal:type_name -> vega.commands.v1.ProtocolUpgradeProposal
+	26, // 22: vega.commands.v1.InputData.oracle_data_submission:type_name -> vega.commands.v1.OracleDataSubmission
+	27, // 23: vega.commands.v1.Transaction.signature:type_name -> vega.commands.v1.Signature
+	0,  // 24: vega.commands.v1.Transaction.version:type_name -> vega.commands.v1.TxVersion
+	3,  // 25: vega.commands.v1.Transaction.pow:type_name -> vega.commands.v1.ProofOfWork
+	26, // [26:26] is the sub-list for method output_type
+	26, // [26:26] is the sub-list for method input_type
+	26, // [26:26] is the sub-list for extension type_name
+	26, // [26:26] is the sub-list for extension extendee
+	0,  // [0:26] is the sub-list for field type_name
 }
 
 func init() { file_vega_commands_v1_transaction_proto_init() }
@@ -951,13 +1006,14 @@ func file_vega_commands_v1_transaction_proto_init() {
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_vega_commands_v1_transaction_proto_rawDesc,
-			NumEnums:      0,
+			NumEnums:      1,
 			NumMessages:   3,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
 		GoTypes:           file_vega_commands_v1_transaction_proto_goTypes,
 		DependencyIndexes: file_vega_commands_v1_transaction_proto_depIdxs,
+		EnumInfos:         file_vega_commands_v1_transaction_proto_enumTypes,
 		MessageInfos:      file_vega_commands_v1_transaction_proto_msgTypes,
 	}.Build()
 	File_vega_commands_v1_transaction_proto = out.File